// frontend/src/api/model/diagnosisCodeCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosisLevel } from './diagnosisLevel';

export interface DiagnosisCodeCreate {
  /**
     * @minLength 1
     * @maxLength 10
     */
  code: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  description: string;
  level?: DiagnosisLevel;
  parent_code?: string | null;
  chapter_code?: string | null;
  is_active?: boolean;
  notes?: string | null;
}